+++ CHANGELOG.md
@@ -9,6 +9,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased] - ReleaseDate
 
+### Changed
+
+- Complete parsing logic for FQNs to match with the Rust spec.
+
 ## [0.1.1] - 2021-07-30
 
 ### Changed